Update complete_multipart tests
The complete_multipart tests tested Ceph behavior implemented in https://github.com/ceph/ceph/pull/40594. The complete_multipart operation was performed twice. The second call doesn't throw an error on Ceph, but it is supposed to throw 404 in S3. I have updated the tests to remove the second call to complete_multipart.